Complete part 6
This was done in a piecemeal way. I discovered the key by looking
at the lowest for normalised hamming distance scores by key length and
found that the 29 gave a meaningful key by looking at the most likely
character for each key belonging to the same transposed group. Using
this key I was able to decrypt the text